% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowprofilePro.R
\docType{data}
\name{codes_pro}
\alias{codes_pro}
\title{PRO field code lookup}
\format{
An object of class \code{list} of length 20.
}
\usage{
codes_pro
}
\value{
A named list of character codes.
}
\description{
Named list of SNOWPACK PRO field codes used by \link{snowprofilePro} and \link{writePro}.
}
\keyword{datasets}
