% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowprofileInstabilitySigns.R
\name{is.snowprofileInstabilitySigns}
\alias{is.snowprofileInstabilitySigns}
\title{Check class snowprofileInstabilitySigns}
\usage{
is.snowprofileInstabilitySigns(x)
}
\arguments{
\item{x}{object to test}
}
\value{
boolean
}
\description{
Check if object is of class \link{snowprofileInstabilitySigns}
}
