% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowprofileTests.R
\name{is.snowprofileTests}
\alias{is.snowprofileTests}
\title{Check class snowprofileTests}
\usage{
is.snowprofileTests(x)
}
\arguments{
\item{x}{object to test}
}
\value{
boolean
}
\description{
Check if object is of class \link{snowprofileTests}
}
