% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sd_sample_uncorrected}
\alias{sd_sample_uncorrected}
\title{fast uncorrected sample standard deviation
https://en.wikipedia.org/wiki/Standard_deviation#Rapid_calculation_methods}
\usage{
sd_sample_uncorrected(x, xbar = mean(x), na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{xbar}{arithmetic mean of x}

\item{na.rm}{remove any NAs before computation of standard deviation?}
}
\value{
uncorrected sample standard deviation (i.e., a numeric scalar)
}
\description{
fast uncorrected sample standard deviation
https://en.wikipedia.org/wiki/Standard_deviation#Rapid_calculation_methods
}
\author{
fherla
}
