% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-SPgroup.R
\docType{data}
\name{SPgroup}
\alias{SPgroup}
\title{Example group of snowprofiles from a mountain drainage}
\format{
A list with 12 entries, that are of class \link{snowprofile}
}
\usage{
SPgroup
}
\description{
A list of 12 snowprofile objects.
}
\examples{

plot(SPgroup, SortMethod = 'unsorted', xticklabels = "originalIndices")
plot(SPgroup, SortMethod = 'hs', xticklabels = "originalIndices")

}
\seealso{
\link{SPpairs}, \link{SPtimeline}, \link{plot.snowprofileSet}
}
\keyword{object}
\keyword{snowprofile}
