% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-SPtimeline_3hourly.R
\docType{data}
\name{SPtimeline_3hourly}
\alias{SPtimeline_3hourly}
\title{Timeseries of snowprofiles in 3 hour resolution
#'}
\format{
A list with several entries, that are of class \link{snowprofile}
}
\usage{
SPtimeline_3hourly
}
\description{
Timeseries of snowprofiles in 3 hour resolution
#'
}
\examples{

summary(SPtimeline_3hourly)
plot(SPtimeline_3hourly)
}
\seealso{
\link{SPtimeline}
}
\keyword{snowprofiles}
