% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.snowprofile.R
\name{summary.snowprofile}
\alias{summary.snowprofile}
\title{Summary of a single snowprofile}
\usage{
\method{summary}{snowprofile}(object, fast = FALSE, ...)
}
\arguments{
\item{object}{snowprofile object}

\item{fast}{boolean switch for twice as fast computation. downside: keep only length-1 meta data, i.e., discard latlon, or nlayers..}

\item{...}{additional arguments for generic method}
}
\value{
data.frame
}
\description{
Summary of a single snowprofile
}
\details{
Creates a one row data.frame where each column contains metadata.

Metadata is determines as elements of the snowprofile object list that are length = 1. An exception is made for latlon where separate columns for lat and lon are produces.

A derived value \code{nLayers} is derived by counting the number of rows in $layers.
}
\examples{

Profile <- SPgroup[[1]]
names(Profile)
summary(Profile)
lapply(SPgroup, summary)

}
\seealso{
\link{summary.snowprofileSet}
}
\author{
shorton
}
